+++ interface/src/components/inputs/index.ts
@@ -1,3 +1,4 @@
 export { default as BlockFormControlLabel } from './BlockFormControlLabel';
 export { default as ValidatedPasswordField } from './ValidatedPasswordField';
 export { default as ValidatedTextField } from './ValidatedTextField';
+export { default as LanguageSelector } from './LanguageSelector';
